DOP-3990: Clean up and consolidate shared logic for both plugins. (#900)
