+++ plugins/gatsby-source-snooty-preview/gatsby-node.js
@@ -4,11 +4,11 @@ const stream = require('stream');
 const { promisify } = require('util');
 const fs = require('fs').promises;
 const { transformBreadcrumbs } = require('../../src/utils/setup/transform-breadcrumbs.js');
-const { saveStaticFiles } = require('../../src/utils/setup/save-asset-files');
-const { getNestedValue } = require('../../src/utils/get-nested-value');
 const pipeline = promisify(stream.pipeline);
 const got = require(`got`);
 const { parser } = require(`stream-json/jsonl/Parser`);
+const { saveStaticFiles } = require('../../src/utils/setup/save-asset-files');
+const { getNestedValue } = require('../../src/utils/get-nested-value');
 const { sourceNodes } = require(`./other-things-to-source`);
 const { fetchClientAccessToken } = require('./utils/kanopy-auth.js');
 const { callPostBuildWebhook } = require('./utils/post-build.js');
