+++ plugins/utils/openapi.js
@@ -1,0 +1,97 @@
+const atlasAdminChangelogS3Prefix = 'https://mongodb-mms-prod-build-server.s3.amazonaws.com/openapi/changelog';
+
+const fetchChangelogData = async (runId, versions) => {
+  try {
+    /* Using metadata runId, fetch OpenAPI Changelog full change list */
+    const changelogResp = await fetch(`${atlasAdminChangelogS3Prefix}/${runId}/changelog.json`);
+    const changelog = await changelogResp.json();
+
+    /* Aggregate all Resources in changelog for frontend filter */
+    const resourcesListSet = new Set();
+    changelog.forEach((release) =>
+      release.paths.forEach(({ httpMethod, path }) => resourcesListSet.add(`${httpMethod} ${path}`))
+    );
+    const changelogResourcesList = Array.from(resourcesListSet);
+
+    /* Fetch most recent Resource Versions' diff */
+    const mostRecentResourceVersions = versions.slice(-2);
+    const mostRecentDiffLabel = mostRecentResourceVersions.join('_');
+    const mostRecentDiffResp = await fetch(`${atlasAdminChangelogS3Prefix}/${runId}/${mostRecentDiffLabel}.json`);
+    const mostRecentDiffData = await mostRecentDiffResp.json();
+
+    return {
+      changelog,
+      changelogResourcesList,
+      mostRecentDiff: {
+        mostRecentDiffLabel,
+        mostRecentDiffData,
+      },
+    };
+  } catch (error) {
+    console.warn('Changelog error: Most recent runId not successful. Using last successful runId to build Changelog.');
+    throw error;
+  }
+};
+
+/* Creates node for ChangelogData, cuyrrently only used for OpenAPI Changelog in cloud-docs. */
+const createOpenAPIChangelogNode = async ({ createNode, createNodeId, createContentDigest, db }) => {
+  try {
+    /* Fetch OpenAPI Changelog metadata */
+    const indexResp = await fetch(`${atlasAdminChangelogS3Prefix}/prod.json`);
+    const index = await indexResp.json();
+
+    const { runId, versions } = index;
+
+    if (!runId || typeof runId !== 'string')
+      throw new Error('OpenAPI Changelog Error: `runId` not available in S3 index.json!');
+
+    let changelogData = {
+      index,
+    };
+    try {
+      const receivedChangelogData = await fetchChangelogData(runId, versions);
+      changelogData = { ...changelogData, ...receivedChangelogData };
+    } catch (error) {
+      /* If any error occurs, fetch last successful metadata and build changelog node */
+      const lastSuccessfulIndex = await db.realmInterface.fetchDocument(
+        'openapi_changelog',
+        'atlas_admin_metadata',
+        {}
+      );
+      const { runId: lastRunId, versions: lastVersions } = lastSuccessfulIndex;
+      const receivedChangelogData = fetchChangelogData(lastRunId, lastVersions);
+      changelogData = { index: lastSuccessfulIndex, ...receivedChangelogData };
+    }
+
+    /* Create Node for useStaticQuery with all Changelog data */
+    createNode({
+      children: [],
+      id: createNodeId('changelogData'),
+      internal: {
+        contentDigest: createContentDigest(changelogData),
+        type: 'ChangelogData',
+      },
+      parent: null,
+      changelogData: changelogData,
+    });
+  } catch (e) {
+    console.error('Error while fetching OpenAPI Changelog data from S3');
+    console.error(e);
+
+    /* Create empty Node for useStaticQuery to ensure successful build */
+    createNode({
+      children: [],
+      id: createNodeId('changelogData'),
+      internal: {
+        contentDigest: createContentDigest({}),
+        type: 'ChangelogData',
+      },
+      parent: null,
+      changelogData: {},
+    });
+  }
+};
+
+module.exports = {
+  createOpenAPIChangelogNode,
+};
